Add files via upload
Country A and country B are both producers and consumers of a good. In the model below I want to evaluate the impact of free trade on the welfare of consumers and producers of both countries and, how that changed after the introduction of tariffs on the importing country. In this example, we tested the effects of Ad Valorem and Specific tariffs.